+++ .github/workflows/python-app.yml
@@ -26,8 +26,8 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install flake8 pytest
-        if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+        pip install flake8 pytest poetry
+        python -m poetry install .
     - name: Lint with flake8
       run: |
         # stop the build if there are Python syntax errors or undefined names
