+++ src/kubeconfig_switcher.egg-info/PKG-INFO
@@ -1,0 +1,13 @@
+Metadata-Version: 2.1
+Name: kubeconfig-switcher
+Version: 0.0.0
+Summary: An application to switch kube config files
+Author: Matthew Post
+License: MIT
+Classifier: Programming Language :: Python :: 3
+Requires-Python: >=3.9
+Description-Content-Type: text/markdown
+License-File: LICENSE
+
+# KubeconfigSwitcher
+A simple way of controlling which Kubenetes cluster you're configured for
