+++ src/kubeconfig_switcher.egg-info/SOURCES.txt
@@ -1,0 +1,18 @@
+.gitignore
+LICENSE
+README.md
+pyproject.toml
+.github/workflows/python-app.yml
+src/__init__.py
+src/kubeconfig_switcher/__init__.py
+src/kubeconfig_switcher/__main__.py
+src/kubeconfig_switcher/app.py
+src/kubeconfig_switcher/data_models.py
+src/kubeconfig_switcher/file_handler.py
+src/kubeconfig_switcher.egg-info/PKG-INFO
+src/kubeconfig_switcher.egg-info/SOURCES.txt
+src/kubeconfig_switcher.egg-info/dependency_links.txt
+src/kubeconfig_switcher.egg-info/requires.txt
+src/kubeconfig_switcher.egg-info/top_level.txt
+tests/__init__.py
+tests/test_file_handler.py
