+++ src/kubeconfig_switcher.egg-info/dependency_links.txt
@@ -1,0 +1,1 @@
+
