+++ src/kubeconfig_switcher.egg-info/requires.txt
@@ -1,0 +1,5 @@
+requests
+pydantic
+
+[:python_version < "3.9"]
+importlib-metadata
