+++ src/kubeconfig_switcher.egg-info/top_level.txt
@@ -1,0 +1,2 @@
+__init__
+kubeconfig_switcher
