+++ src/kubeconfig_switcher/__main__.py
